Update main.yml workflow to install additional dependencies and improve ISO creation process
The commit updates the main.yml workflow file by adding the installation of syslinux-utils as a dependency. It also improves the ISO creation process by specifying additional options for xorriso, such as setting the boot catalog, boot image, and boot information table. These changes aim to enhance the overall functionality and performance of the workflow.